Added assimp library, fixed application static initialization,

+++ Graphics/src/GraphicsInputLayout.cpp
@@ -52,7 +52,7 @@ namespace pawn {
 	}
 
 	void GraphicsInputLayout::RecalculateStride() {
-		uint32_t offset = 0;
+		size_t offset = 0;
 		m_Stride = 0;
 
 		for (auto& element : m_Elements) {
